+++ src/store/state.js
@@ -12,7 +12,9 @@ export default {
   isLoading: false,
   currentPage: 1,
   pageSize: 10,
-  biobankPagination: {},
+  initialBiobankCount: 0,
+  initialBiobankIds: [],
+  biobankCount: 0,
   isPodium: false,
   podiumCollectionIds: [],
   error: null,
